enum validation
